requirements.txt
botorch==0.7.2
cvxopt==1.3.0
torch==1.12.1  # install the appropriate CUDA version manually if necessary, see: https://pytorch.org/get-started/previous-versions/
rdkit-pypi==2022.03.5
torch_geometric==2.1.0.post1
numpy==1.23.4
pandas==1.5.0
matplotlib==3.6.0
seaborn==0.12.1
pyprojroot==0.2.0
selfies==2.1.1
